# _extract_sle/Store-Link-Engine/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ziadah Product Intelligence Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type StoreStatus = (typeof StoreStatus)[keyof typeof StoreStatus];

export const StoreStatus = {
  pending: "pending",
  synced: "synced",
  analyzing: "analyzing",
  analyzed: "analyzed",
  error: "error",
} as const;

export interface Store {
  id: number;
  name: string;
  url: string;
  /** @nullable */
  platform: string | null;
  status: StoreStatus;
  productCount: number;
  /** @nullable */
  lastSyncedAt: string | null;
  /** @nullable */
  lastAnalyzedAt: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateStoreBody {
  name: string;
  url: string;
}

export interface UpdateStoreBody {
  name?: string;
  url?: string;
}

/**
 * @nullable
 */
export type ProductRole = (typeof ProductRole)[keyof typeof ProductRole] | null;

export const ProductRole = {
  main: "main",
  cross_sell: "cross_sell",
  upsell: "upsell",
} as const;

export interface Product {
  id: number;
  storeId: number;
  /** @nullable */
  externalId: string | null;
  title: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  price: number | null;
  /** @nullable */
  compareAtPrice: number | null;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  productUrl: string | null;
  /** @nullable */
  category: string | null;
  /** @nullable */
  tags: string | null;
  inStock: boolean;
  /** @nullable */
  role: ProductRole;
  createdAt: string;
  updatedAt: string;
}

export interface SyncResult {
  synced: number;
  total: number;
  platform: string;
  message: string;
}

export interface ProductRecommendation {
  productId: number;
  title: string;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  price: number | null;
  reason: string;
}

export interface AnalysisResult {
  storeId: number;
  analyzedAt: string;
  mainProduct: ProductRecommendation;
  crossSells: ProductRecommendation[];
  upsells: ProductRecommendation[];
  summary: string;
}

export interface RecommendationsSummary {
  storeId: number;
  storeName: string;
  totalProducts: number;
  /** @nullable */
  mainProductTitle: string | null;
  crossSellCount: number;
  upsellCount: number;
  /** @nullable */
  lastAnalyzedAt: string | null;
  status: string;
}