# _extract_sle/Store-Link-Engine/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ziadah Product Intelligence Engine API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all registered stores
 */
export const ListStoresResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  url: zod.string(),
  platform: zod.string().nullable(),
  status: zod.enum(["pending", "synced", "analyzing", "analyzed", "error"]),
  productCount: zod.number(),
  lastSyncedAt: zod.string().nullable(),
  lastAnalyzedAt: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListStoresResponse = zod.array(ListStoresResponseItem);

/**
 * @summary Register a new store
 */
export const CreateStoreBody = zod.object({
  name: zod.string(),
  url: zod.string(),
});

/**
 * @summary Get a single store
 */
export const GetStoreParams = zod.object({
  id: zod.coerce.number(),
});

export const GetStoreResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  url: zod.string(),
  platform: zod.string().nullable(),
  status: zod.enum(["pending", "synced", "analyzing", "analyzed", "error"]),
  productCount: zod.number(),
  lastSyncedAt: zod.string().nullable(),
  lastAnalyzedAt: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a store
 */
export const UpdateStoreParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateStoreBody = zod.object({
  name: zod.string().optional(),
  url: zod.string().optional(),
});

export const UpdateStoreResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  url: zod.string(),
  platform: zod.string().nullable(),
  status: zod.enum(["pending", "synced", "analyzing", "analyzed", "error"]),
  productCount: zod.number(),
  lastSyncedAt: zod.string().nullable(),
  lastAnalyzedAt: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a store
 */
export const DeleteStoreParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Fetch and sync products from the store URL
 */
export const SyncStoreParams = zod.object({
  id: zod.coerce.number(),
});

export const SyncStoreResponse = zod.object({
  synced: zod.number(),
  total: zod.number(),
  platform: zod.string(),
  message: zod.string(),
});

/**
 * @summary Run AI analysis to identify main product, cross-sells and upsells
 */
export const AnalyzeStoreParams = zod.object({
  id: zod.coerce.number(),
});

export const AnalyzeStoreResponse = zod.object({
  storeId: zod.number(),
  analyzedAt: zod.string(),
  mainProduct: zod.object({
    productId: zod.number(),
    title: zod.string(),
    imageUrl: zod.string().nullable(),
    price: zod.number().nullable(),
    reason: zod.string(),
  }),
  crossSells: zod.array(
    zod.object({
      productId: zod.number(),
      title: zod.string(),
      imageUrl: zod.string().nullable(),
      price: zod.number().nullable(),
      reason: zod.string(),
    }),
  ),
  upsells: zod.array(
    zod.object({
      productId: zod.number(),
      title: zod.string(),
      imageUrl: zod.string().nullable(),
      price: zod.number().nullable(),
      reason: zod.string(),
    }),
  ),
  summary: zod.string(),
});

/**
 * @summary List all products for a store
 */
export const ListStoreProductsParams = zod.object({
  id: zod.coerce.number(),
});

export const ListStoreProductsResponseItem = zod.object({
  id: zod.number(),
  storeId: zod.number(),
  externalId: zod.string().nullable(),
  title: zod.string(),
  description: zod.string().nullable(),
  price: zod.number().nullable(),
  compareAtPrice: zod.number().nullable(),
  imageUrl: zod.string().nullable(),
  productUrl: zod.string().nullable(),
  category: zod.string().nullable(),
  tags: zod.string().nullable(),
  inStock: zod.boolean(),
  role: zod
    .union([
      zod.literal("main"),
      zod.literal("cross_sell"),
      zod.literal("upsell"),
      zod.literal(null),
    ])
    .nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListStoreProductsResponse = zod.array(
  ListStoreProductsResponseItem,
);

/**
 * @summary Get the latest analysis results for a store
 */
export const GetLatestAnalysisParams = zod.object({
  id: zod.coerce.number(),
});

export const GetLatestAnalysisResponse = zod.object({
  storeId: zod.number(),
  analyzedAt: zod.string(),
  mainProduct: zod.object({
    productId: zod.number(),
    title: zod.string(),
    imageUrl: zod.string().nullable(),
    price: zod.number().nullable(),
    reason: zod.string(),
  }),
  crossSells: zod.array(
    zod.object({
      productId: zod.number(),
      title: zod.string(),
      imageUrl: zod.string().nullable(),
      price: zod.number().nullable(),
      reason: zod.string(),
    }),
  ),
  upsells: zod.array(
    zod.object({
      productId: zod.number(),
      title: zod.string(),
      imageUrl: zod.string().nullable(),
      price: zod.number().nullable(),
      reason: zod.string(),
    }),
  ),
  summary: zod.string(),
});

/**
 * @summary Get product recommendations summary for a store
 */
export const GetRecommendationsParams = zod.object({
  id: zod.coerce.number(),
});

export const GetRecommendationsResponse = zod.object({
  storeId: zod.number(),
  storeName: zod.string(),
  totalProducts: zod.number(),
  mainProductTitle: zod.string().nullable(),
  crossSellCount: zod.number(),
  upsellCount: zod.number(),
  lastAnalyzedAt: zod.string().nullable(),
  status: zod.string(),
});