# _extract_sle/Store-Link-Engine/lib/api-zod/src/generated/types/createStoreBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ziadah Product Intelligence Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateStoreBody {
  name: string;
  url: string;
}