# _extract_sle/Store-Link-Engine/lib/db/src/schema/products.ts
import { pgTable, text, serial, timestamp, integer, real, boolean } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { storesTable } from "./stores";

export const productsTable = pgTable("products", {
  id: serial("id").primaryKey(),
  storeId: integer("store_id").notNull().references(() => storesTable.id, { onDelete: "cascade" }),
  externalId: text("external_id"),
  title: text("title").notNull(),
  description: text("description"),
  price: real("price"),
  compareAtPrice: real("compare_at_price"),
  imageUrl: text("image_url"),
  productUrl: text("product_url"),
  category: text("category"),
  tags: text("tags"),
  inStock: boolean("in_stock").notNull().default(true),
  role: text("role"),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertProductSchema = createInsertSchema(productsTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertProduct = z.infer<typeof insertProductSchema>;
export type Product = typeof productsTable.$inferSelect;