# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ziadah Product Intelligence Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductRecommendation } from "./productRecommendation";

export interface AnalysisResult {
  storeId: number;
  analyzedAt: string;
  mainProduct: ProductRecommendation;
  crossSells: ProductRecommendation[];
  upsells: ProductRecommendation[];
  summary: string;
}