# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ziadah Product Intelligence Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductRole } from "./productRole";

export interface Product {
  id: number;
  storeId: number;
  /** @nullable */
  externalId: string | null;
  title: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  price: number | null;
  /** @nullable */
  compareAtPrice: number | null;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  productUrl: string | null;
  /** @nullable */
  category: string | null;
  /** @nullable */
  tags: string | null;
  inStock: boolean;
  /** @nullable */
  role: ProductRole;
  createdAt: string;
  updatedAt: string;
}